fixed navigation to logged_in
