+++ android/app/src/main/AndroidManifest.xml
@@ -38,14 +38,14 @@
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-            <!--<intent-filter>-->
-                <!--<data android:scheme="@string/APP_SCHEME" />-->
+            <intent-filter>
+                <data android:scheme="@string/APP_SCHEME" />
 
-                <!--<action android:name="android.intent.action.VIEW" />-->
+                <action android:name="android.intent.action.VIEW" />
 
-                <!--<category android:name="android.intent.category.DEFAULT" />-->
-                <!--<category android:name="android.intent.category.BROWSABLE" />-->
-            <!--</intent-filter>-->
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
+            </intent-filter>
             <intent-filter android:autoVerify="true">
                 <action android:name="android.intent.action.VIEW" />
