+++ android/app/src/main/res/menu/my_info_menu.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@id/save"
+        android:id="@+id/save"
         android:orderInCategory="1"
         android:title="@string/PROFILE_MY_INFO_SAVE"
         app:showAsAction="always|withText" />
