--- android/common/src/main/res/values/ids.xml
@@ -1,112 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <item name="marketing_screen_content" type="id" />
-    <item name="chatFragment" type="id"/>
-    <item name="logo_content" type="id" />
-    <item name="profile_screen_content" type="id" />
-    <item name="profile_member_info_screen_content" type="id" />
-    <item name="getHedvig" type="id" />
-    <item name="sayHello" type="id" />
-    <item name="pager" type="id" />
-    <item name="blurOverlay" type="id" />
-    <item name="marketing_hedvig_logo" type="id" />
-    <item name="storyProgressIndicatorContainer" type="id" />
-    <item name="hedvigFaceAnimation" type="id" />
-    <item name="login" type="id" />
-    <item name="activity_marketing" type="id" />
-    <item name="loadingSpinner" type="id" />
-    <item name="profile_charity_image" type="id" />
-    <item name="myInfoRow" type="id" />
-    <item name="myHomeRow" type="id" />
-    <item name="rowContainer" type="id" />
-    <item name="profile_my_info_icon" type="id" />
-    <item name="profile_my_info_header" type="id" />
-    <item name="sphereContainer" type="id" />
-    <item name="emailInput" type="id" />
-    <item name="phoneNumberInput" type="id" />
-    <item name="contactDetailsContainer" type="id" />
-    <item name="paymentRow" type="id" />
-    <item name="profile_menu_row_description" type="id" />
-    <item name="profile_payment_screen_content" type="id" />
-    <item name="profile_payment_header" type="id" />
-    <item name="sphere" type="id" />
-    <item name="priceSphere" type="id" />
-    <item name="paymentDetailsContainer" type="id" />
-    <item name="bankContainer" type="id" />
-    <item name="bankName" type="id" />
-    <item name="accountNumber" type="id" />
-    <item name="changeBankAccount" type="id" />
-    <item name="profile_payment_deductible_sphere" type="id" />
-    <item name="toolbar" type="id" />
-    <item name="profile_menu_row" type="id" />
-    <item name="profile_menu_row_icon" type="id" />
-    <item name="profile_menu_row_name" type="id" />
-    <item name="profile_menu_item_description" type="id" />
-    <item name="profileInfoRow" type="id" />
-    <item name="profilePaymentRow" type="id" />
-    <item name="name" type="id" />
-    <item name="app_bar" type="id" />
-    <item name="collapsingToolbar" type="id" />
-    <item name="save" type="id" />
-    <item name="coinsuredRow" type="id" />
-    <item name="charityRow" type="id" />
-    <item name="insuranceCertificateRow" type="id" />
-    <item name="feedbackRow" type="id" />
-    <item name="logout" type="id" />
-    <item name="address" type="id" />
-    <item name="postalNumber" type="id" />
-    <item name="insuranceType" type="id" />
-    <item name="infoContainer" type="id" />
-    <item name="pdf" type="id" />
-    <item name="bugReportEmail" type="id" />
-    <item name="playLink" type="id" />
-    <item name="coinsuredSphere" type="id" />
-    <item name="textContainer" type="id" />
-    <item name="sphereText" type="id" />
-    <item name="selectedCharityContainer" type="id" />
-    <item name="selectCharityContainer" type="id" />
-    <item name="selectedCharityBanner" type="id" />
-    <item name="selectedCharityCardTitle" type="id" />
-    <item name="selectedCharityCardParagraph" type="id" />
-    <item name="charityIcon" type="id" />
-    <item name="selectCharityExplainer" type="id" />
-    <item name="cashbackOptionTitle" type="id" />
-    <item name="cashbackOptionParagraph" type="id" />
-    <item name="cashbackOptions" type="id" />
-    <item name="cashbackSelect" type="id" />
-    <item name="separator" type="id" />
-    <item name="loadingAnimation" type="id" />
-    <item name="aboutAppRow" type="id" />
-    <item name="licenseAttributions" type="id" />
-    <item name="versionNumber" type="id" />
-    <item name="memberId" type="id" />
-    <item name="changeHomeInformation" type="id" />
-    <item name="dialogTitle" type="id" />
-    <item name="dialogConfirm" type="id" />
-    <item name="dialogCancel" type="id" />
-    <item name="dialogParagraph" type="id" />
-    <item name="trustlyContainer" type="id" />
-    <item name="successScreen" type="id" />
-    <item name="failScreen" type="id" />
-    <item name="successIcon" type="id" />
-    <item name="successTitle" type="id" />
-    <item name="successParagraph" type="id" />
-    <item name="failureIcon" type="id" />
-    <item name="failureTitle" type="id" />
-    <item name="failureParagraph" type="id" />
-    <item name="trustlyFailClose" type="id" />
-    <item name="trustlySuccessClose" type="id" />
-    <item name="deductibleSphere" type="id" />
-    <item name="autogiroDate" type="id" />
-    <item name="cardTitle" type="id" />
-    <item name="cardParagraph" type="id" />
-    <item name="cardSeparator" type="id" />
-    <item name="connectBankAccount" type="id" />
-    <item name="connectBankAccountContainer" type="id" />
-    <item name="licenseList" type="id" />
-    <item name="emailInputContainer" type="id" />
-    <item name="phoneNumberInputContainer" type="id" />
-    <item name="bankAccountUnderChangeParagraph" type="id" />
-    <item name="profileReferralRow" type="id" />
-    <item name="iconNavigateNext" type="id" />
-</resources>
