+++ android/logged_in/src/main/res/layout/cashback_option.xml
@@ -11,7 +11,7 @@
         android:layout_height="match_parent">
 
         <TextView
-            android:id="@id/cashbackOptionTitle"
+            android:id="@+id/cashbackOptionTitle"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="20dp"
@@ -22,7 +22,7 @@
             tools:text="SOS Barnbyar" />
 
         <TextView
-            android:id="@id/cashbackOptionParagraph"
+            android:id="@+id/cashbackOptionParagraph"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="20dp"
@@ -30,19 +30,19 @@
             android:layout_marginEnd="20dp"
             android:theme="@style/Body.Small"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/cashbackOptionTitle"
+            app:layout_constraintTop_toBottomOf="@+id/cashbackOptionTitle"
             tools:text="För att bli en trygg vuxen krävs en trygg barndom. Därför arbetar SOS Barnbyar med långsiktiga insatser för att inte ett enda barn ska behöva växa upp ensamt." />
 
         <View
-            android:id="@id/separator"
+            android:id="@+id/separator"
             android:layout_width="match_parent"
             android:layout_height="1dp"
             android:layout_marginTop="24dp"
             android:background="@color/light_gray"
-            app:layout_constraintTop_toBottomOf="@id/cashbackOptionParagraph" />
+            app:layout_constraintTop_toBottomOf="@+id/cashbackOptionParagraph" />
 
         <Button
-            android:id="@id/cashbackSelect"
+            android:id="@+id/cashbackSelect"
             style="?android:attr/borderlessButtonStyle"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -51,7 +51,7 @@
             android:theme="@style/DialogButton"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/separator"
+            app:layout_constraintTop_toBottomOf="@+id/separator"
             tools:text="Välj SOS Barnbyar" />
 
     </android.support.constraint.ConstraintLayout>
