+++ android/logged_in/src/main/res/layout/common_claim_first_message.xml
@@ -46,7 +46,7 @@
         android:textColor="@color/off_white"
         android:textSize="@dimen/text_medium"
         app:backgroundTint="@color/purple"
-        app:layout_constraintBottom_toTopOf="@id/login"
+        app:layout_constraintBottom_toTopOf="@+id/login"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         tools:backgroundTint="@color/purple"
