+++ android/logged_in/src/main/res/layout/fragment_about_app.xml
@@ -20,14 +20,14 @@
             android:orientation="vertical">
 
             <com.hedvig.logged_in.profile.ui.ProfileMenuRow
-                android:id="@id/licenseAttributions"
+                android:id="@+id/licenseAttributions"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:theme="@style/Body.Small"
                 app:name="@string/PROFILE_ABOUT_APP_LICENSE_ATTRIBUTIONS"/>
 
             <TextView
-                android:id="@id/versionNumber"
+                android:id="@+id/versionNumber"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:gravity="center"
@@ -35,7 +35,7 @@
                 tools:text="Version: 3.4.5"/>
 
             <TextView
-                android:id="@id/memberId"
+                android:id="@+id/memberId"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:gravity="center"
