+++ android/logged_in/src/main/res/layout/fragment_feedback.xml
@@ -51,7 +51,7 @@
                 android:theme="@style/ProfileFieldLabel"/>
 
             <TextView
-                android:id="@id/bugReportEmail"
+                android:id="@+id/bugReportEmail"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="16dp"
@@ -70,7 +70,7 @@
                 android:theme="@style/ProfileFieldLabel"/>
 
             <TextView
-                android:id="@id/playLink"
+                android:id="@+id/playLink"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="16dp"
