+++ android/logged_in/src/main/res/layout/fragment_my_info.xml
@@ -23,11 +23,11 @@
             <include layout="@layout/loading_spinner" />
 
             <include
-                android:id="@id/sphereContainer"
+                android:id="@+id/sphereContainer"
                 layout="@layout/sphere_container" />
 
             <LinearLayout
-                android:id="@id/contactDetailsContainer"
+                android:id="@+id/contactDetailsContainer"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:focusable="true"
@@ -35,14 +35,14 @@
                 android:orientation="vertical"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/sphereContainer">
+                app:layout_constraintTop_toBottomOf="@+id/sphereContainer">
 
                 <TextView
                     style="@style/SectionHeader"
                     android:text="@string/PROFILE_MY_INFO_SECTION_TITLE_CONTACT_DETAILS" />
 
                 <android.support.design.widget.TextInputLayout
-                    android:id="@id/emailInputContainer"
+                    android:id="@+id/emailInputContainer"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="12dp"
@@ -52,7 +52,7 @@
                     app:errorTextAppearance="@style/ProfileField.ValidationError">
 
                     <android.support.design.widget.TextInputEditText
-                        android:id="@id/emailInput"
+                        android:id="@+id/emailInput"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
                         android:hint="@string/PROFILE_MY_INFO_EMAIL_LABEL"
@@ -63,7 +63,7 @@
 
 
                 <android.support.design.widget.TextInputLayout
-                    android:id="@id/phoneNumberInputContainer"
+                    android:id="@+id/phoneNumberInputContainer"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="12dp"
@@ -73,7 +73,7 @@
                     app:errorTextAppearance="@style/ProfileField.ValidationError">
 
                     <android.support.design.widget.TextInputEditText
-                        android:id="@id/phoneNumberInput"
+                        android:id="@+id/phoneNumberInput"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
                         android:hint="@string/PROFILE_MY_INFO_PHONE_NUMBER_LABEL"
