+++ android/logged_in/src/main/res/layout/fragment_payment.xml
@@ -22,7 +22,7 @@
             <include layout="@layout/loading_spinner"/>
 
             <android.support.constraint.ConstraintLayout
-                android:id="@id/sphereContainer"
+                android:id="@+id/sphereContainer"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
@@ -33,7 +33,7 @@
                 tools:visibility="visible">
 
                 <ImageView
-                    android:id="@id/priceSphere"
+                    android:id="@+id/priceSphere"
                     android:layout_width="@dimen/profile_sphere_side_length"
                     android:layout_height="@dimen/profile_sphere_side_length"
                     android:layout_alignParentStart="true"
@@ -70,7 +70,7 @@
                     app:layout_constraintStart_toStartOf="parent">
 
                     <ImageView
-                        android:id="@id/deductibleSphere"
+                        android:id="@+id/deductibleSphere"
                         android:layout_width="80dp"
                         android:layout_height="80dp"
                         android:layout_alignParentStart="true"
@@ -103,14 +103,14 @@
             </android.support.constraint.ConstraintLayout>
 
             <LinearLayout
-                android:id="@id/paymentDetailsContainer"
+                android:id="@+id/paymentDetailsContainer"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:orientation="vertical"
                 android:visibility="gone"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/sphereContainer"
+                app:layout_constraintTop_toBottomOf="@+id/sphereContainer"
                 tools:visibility="visible">
 
                 <TextView
@@ -134,7 +134,7 @@
                     android:textSize="14sp"/>
 
                 <TextView
-                    android:id="@id/autogiroDate"
+                    android:id="@+id/autogiroDate"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="24dp"
@@ -155,7 +155,7 @@
                     android:textSize="16sp"/>
 
                 <TextView
-                    android:id="@id/bankName"
+                    android:id="@+id/bankName"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="24dp"
@@ -166,7 +166,7 @@
                     tools:text="SEB"/>
 
                 <TextView
-                    android:id="@id/accountNumber"
+                    android:id="@+id/accountNumber"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="24dp"
@@ -177,7 +177,7 @@
                     tools:text="**** 8347"/>
 
                 <TextView
-                    android:id="@id/bankAccountUnderChangeParagraph"
+                    android:id="@+id/bankAccountUnderChangeParagraph"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="24dp"
@@ -192,7 +192,7 @@
             </LinearLayout>
 
             <android.support.design.card.MaterialCardView
-                android:id="@id/connectBankAccountContainer"
+                android:id="@+id/connectBankAccountContainer"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="8dp"
@@ -201,7 +201,7 @@
                 android:visibility="gone"
                 app:cardCornerRadius="8dp"
                 app:cardUseCompatPadding="true"
-                app:layout_constraintTop_toBottomOf="@id/sphereContainer">
+                app:layout_constraintTop_toBottomOf="@+id/sphereContainer">
 
                 <android.support.constraint.ConstraintLayout
                     android:layout_width="match_parent"
@@ -209,7 +209,7 @@
                     android:orientation="vertical">
 
                     <TextView
-                        android:id="@id/cardTitle"
+                        android:id="@+id/cardTitle"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_marginStart="20dp"
@@ -220,7 +220,7 @@
                         app:layout_constraintTop_toTopOf="parent"/>
 
                     <TextView
-                        android:id="@id/cardParagraph"
+                        android:id="@+id/cardParagraph"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
                         android:layout_marginStart="20dp"
@@ -228,18 +228,18 @@
                         android:layout_marginEnd="20dp"
                         android:text="@string/PROFILE_PAYMENT_CONNECT_DIRECT_DEBIT_DESCRIPTION"
                         android:theme="@style/Body.Small"
-                        app:layout_constraintTop_toBottomOf="@id/cardTitle"/>
+                        app:layout_constraintTop_toBottomOf="@+id/cardTitle"/>
 
                     <View
-                        android:id="@id/cardSeparator"
+                        android:id="@+id/cardSeparator"
                         android:layout_width="match_parent"
                         android:layout_height="1dp"
                         android:layout_marginTop="24dp"
                         android:background="@color/light_gray"
-                        app:layout_constraintTop_toBottomOf="@id/cardParagraph"/>
+                        app:layout_constraintTop_toBottomOf="@+id/cardParagraph"/>
 
                     <Button
-                        android:id="@id/connectBankAccount"
+                        android:id="@+id/connectBankAccount"
                         style="?android:attr/borderlessButtonStyle"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
@@ -255,16 +255,16 @@
             </android.support.design.card.MaterialCardView>
 
             <View
-                android:id="@id/separator"
+                android:id="@+id/separator"
                 android:layout_width="match_parent"
                 android:layout_height="1dp"
                 android:layout_marginTop="24dp"
                 android:background="@color/light_gray"
                 android:visibility="gone"
-                app:layout_constraintTop_toBottomOf="@id/paymentDetailsContainer"/>
+                app:layout_constraintTop_toBottomOf="@+id/paymentDetailsContainer"/>
 
             <Button
-                android:id="@id/changeBankAccount"
+                android:id="@+id/changeBankAccount"
                 style="?android:attr/borderlessButtonStyle"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -276,7 +276,7 @@
                 android:visibility="gone"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/separator"/>
+                app:layout_constraintTop_toBottomOf="@+id/separator"/>
         </android.support.constraint.ConstraintLayout>
 
     </android.support.v4.widget.NestedScrollView>
