+++ android/logged_in/src/main/res/layout/sphere_container.xml
@@ -13,7 +13,7 @@
     tools:visibility="visible">
 
     <ImageView
-        android:id="@id/sphere"
+        android:id="@+id/sphere"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:contentDescription="@null"
