+++ android/logged_in/src/main/res/menu/logged_in_menu.xml
