+++ android/onboarding/build.gradle
@@ -10,6 +10,10 @@ android {
         targetSdkVersion 28
         versionCode 1
         versionName "1.0"
+
+        ndk {
+            abiFilters 'armeabi-v7a', 'x86'
+        }
     }
 }
