+++ android/onboarding/src/main/java/com/hedvig/onboarding/OnBoardingActivity.kt
@@ -5,38 +5,74 @@ import android.content.Intent
 import android.os.Build
 import android.os.Bundle
 import androidx.navigation.findNavController
-import com.facebook.react.ReactApplication
+import com.airbnb.android.react.lottie.LottiePackage
+import com.apollographql.apollo.ApolloClient
 import com.facebook.react.ReactInstanceManager
 import com.facebook.react.ReactNativeHost
 import com.facebook.react.modules.core.DefaultHardwareBackBtnHandler
 import com.facebook.react.modules.core.PermissionAwareActivity
 import com.facebook.react.modules.core.PermissionListener
+import com.facebook.react.shell.MainReactPackage
 import com.hedvig.app.BaseActivity
+import com.hedvig.app.BuildConfig
 import com.hedvig.app.service.LoginStatus
 import com.hedvig.common.util.NavigationAnalytics
 import com.hedvig.common.util.extensions.proxyNavigate
+import com.hedvig.onboarding.react.ActivityStarterReactPackage
 import com.hedvig.onboarding.react.AsyncStorageNative
+import com.hedvig.onboarding.react.NativeRoutingPackage
+import com.horcrux.svg.SvgPackage
+import com.leo_pharma.analytics.AnalyticsPackage
+import com.lugg.ReactNativeConfig.ReactNativeConfigPackage
+import com.rnfs.RNFSPackage
+import com.rnim.rn.audio.ReactNativeAudioPackage
+import com.zmxv.RNSound.RNSoundPackage
 import io.branch.rnbranch.RNBranchModule
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.rxkotlin.plusAssign
-import io.reactivex.schedulers.Schedulers
+import io.branch.rnbranch.RNBranchPackage
+import io.sentry.RNSentryPackage
 import org.koin.android.ext.android.inject
 import timber.log.Timber
 
 class OnBoardingActivity : BaseActivity(), DefaultHardwareBackBtnHandler, PermissionAwareActivity {
+
+    private val apolloClient: ApolloClient by inject()
 
     private val navController by lazy { findNavController(R.id.onBoardingNavigationHost) }
 
     private var permissionListener: PermissionListener? = null
 
     val asyncStorageNative: AsyncStorageNative by inject()
-    val loggedInService: com.hedvig.onboarding.LoginStatusService by inject()
+
+    val loggedInService: LoginStatusService by inject()
 
     private val reactInstanceManager: ReactInstanceManager
         get() = reactNativeHost.reactInstanceManager
 
-    private val reactNativeHost: ReactNativeHost
-        get() = (application as ReactApplication).reactNativeHost
+    //todo inject this with koin
+    val reactNativeHost by lazy {
+        Timber.i("application-> ${application == null}")
+        object : ReactNativeHost(application) {
+            override fun getUseDeveloperSupport() = BuildConfig.DEBUG
+
+            override fun getPackages() = listOf(
+                ActivityStarterReactPackage(apolloClient, asyncStorageNative),
+                MainReactPackage(),
+//            ImagePickerPackage(),
+                RNFSPackage(),
+                SvgPackage(),
+                ReactNativeConfigPackage(),
+                RNSoundPackage(),
+                RNSentryPackage(),
+                RNBranchPackage(),
+                ReactNativeAudioPackage(),
+                AnalyticsPackage(),
+                LottiePackage(),
+                NativeRoutingPackage(apolloClient)
+            )
+
+            override fun getJSMainModuleName() = "index.android"
+        }
+    }
 
 
     @TargetApi(Build.VERSION_CODES.M)
