+++ android/onboarding/src/main/res/layout/onboarding_loading_spinner.xml
